secret was not found-bug

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\BlogPost;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StorePost;
 use Illuminate\Support\Facades\DB;
 
@@ -15,7 +16,7 @@ class PostController extends Controller
     public function __construct()
     {
         $this->middleware('auth')
-        ->only(['create','store','update','destroy']);
+             ->only(['create','store','update','destroy']);
     }
     /**
      * Display a listing of the resource.
@@ -59,6 +60,7 @@ class PostController extends Controller
 
     public function create()
     {
+        // $this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -76,12 +78,24 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update', $post);
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog-post");
+        // }
+
         return view('posts.edit', ['posts'=>$post]);
     }
     public function update(StorePost $request, $id)
     {
 
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize('update', $post);
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog-post");
+        // }
+
         $validatedData = $request->validated();
         // $posts->title=$validatedData['title'];
         // $posts->content=$validatedData['content'];
@@ -94,6 +108,14 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // $this->authorize($post); // jos bolje i jednostavanije
+        $this->authorize('delete', $post);
+        // $this->authorize('posts.delete', $post); //ovo gore lakse generisati policy klase
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog-post");
+        // }
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
